+++ app/routes.php
@@ -9,6 +9,7 @@
  */
 
 use App\Http\Controllers\LoginController;
+use App\Http\Controllers\ProductController;
 use App\Http\Controllers\UserController;
 use Framework\Routing\Router;
 
@@ -22,15 +23,7 @@ Router::addController('login', new LoginController(), 'POST');
 // TODO only if logged in
 // TODO only if is admin
 // TODO middelware / role check
-Router::addResource('user', new UserController());
+Router::addModel('product', new ProductController());
+Router::addModel('user', new UserController());
 
 Router::pathNotFound(function ($path) { echo 'cannot find path "' . $path. '"'; });
-
-// TODO add to documentation
-// Router::addController('user/create', new TestController()); // form for new resource
-// Router::addController('user/store', new TestController()); // store new resource
-// Router::addController('user/edit', new TestController()); // show form for editing
-// Router::addController('user/update', new TestController()); // store the edit changes
-// Router::addController('user/destroy', new TestController()); // remove resource
-// Router::addController('user/index', new TestController()); // List the resources
-// index(), create(), store($model), show($id), edit($id), update($request, id), destory(id)