+++ resources/lang/de.php
@@ -1,9 +1,16 @@
 <?php
 
 return [
+    'Create' => 'Erstellen',
+    'CreateNewProduct' => 'Neues Produkt erstellen',
+    'Delete' => 'Löschen',
     'Developer' => 'Entwickler',
     'Home' => 'Startseite',
+    'IsDiscontinued' => 'Ist abgekündigt',
     'Login' => 'Anmelden',
     'Password' => 'Passwort',
+    'Product' => 'Produkt',
+    'ProductName' => 'Produktname',
+    'Save' => 'Speichern',
     'Username' => 'Benutzername',
 ];