+++ resources/lang/en.php
@@ -1,9 +1,16 @@
 <?php
 
 return [
+    'Create' => 'Create',
+    'CreateNewProduct' => 'Create new product',
+    'Delete' => 'Delete',
     'Developer' => 'Developer',
     'Home' => 'Home',
+    'IsDiscontinued' => 'Is discontinued',
     'Login' => 'Login',
     'Password' => 'Password',
+    'Product' => 'Product',
+    'ProductName' => 'Product name',
+    'Save' => 'Save',
     'Username' => 'Username',
 ];