+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added ConfigReader
 - Added support for database connections and migrations
 - Added command to create a new controller
+- Added all CRUD actions for product with ProductController
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2